+++ Libplanet/Blockchain/BlockChainStates.cs
@@ -7,7 +7,6 @@ using Libplanet.Blocks;
 using Libplanet.Consensus;
 using Libplanet.State;
 using Libplanet.Store;
-using Libplanet.Store.Trie;
 using static Libplanet.Blockchain.KeyConverters;
 
 namespace Libplanet.Blockchain
@@ -136,12 +135,6 @@ namespace Libplanet.Blockchain
             }
 
             throw new IncompleteBlockStatesException(offset);
-        }
-
-        /// <inheritdoc cref="IBlockChainStates.GetTrie"/>
-        public ITrie? GetTrie(BlockHash offset)
-        {
-            return _stateStore.GetStateRoot(_store.GetBlockDigest(offset)?.StateRootHash);
         }
     }
 }